+++ src/main/java/miniprogram/server/ServerApplication.java
@@ -1,13 +1,13 @@
-package com.miniprogram.java;
+package miniprogram.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class ServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
+        SpringApplication.run(ServerApplication.class, args);
     }
 
 }
