+++ src/main/java/miniprogram/server/controller/IndexController.java
@@ -1,4 +1,4 @@
-package com.miniprogram.java.controller;
+package miniprogram.server.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
