+++ src/test/java/miniprogram/server/ServerApplicationTests.java
@@ -1,10 +1,10 @@
-package com.miniprogram.java;
+package miniprogram.server;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class ServerApplicationTests {
 
     @Test
     void contextLoads() {
